Fixed #1926: PostgreSQL: Flyway uses public schema instead of current_schema when search_path starts with $user

+++ flyway-core/src/main/java/org/flywaydb/core/internal/schemahistory/JdbcTableSchemaHistory.java
@@ -24,7 +24,6 @@ import org.flywaydb.core.api.logging.LogFactory;
 import org.flywaydb.core.api.resolver.ResolvedMigration;
 import org.flywaydb.core.internal.database.Connection;
 import org.flywaydb.core.internal.database.Database;
-import org.flywaydb.core.internal.database.Schema;
 import org.flywaydb.core.internal.database.Table;
 import org.flywaydb.core.internal.exception.FlywaySqlException;
 import org.flywaydb.core.internal.util.jdbc.JdbcTemplate;
@@ -40,7 +39,7 @@ import java.util.concurrent.Callable;
 /**
  * Supports reading and writing to the schema history table.
  */
-public class JdbcTableSchemaHistory extends SchemaHistory {
+class JdbcTableSchemaHistory extends SchemaHistory {
     private static final Log LOG = LogFactory.getLog(JdbcTableSchemaHistory.class);
 
     /**
@@ -51,12 +50,7 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
     /**
      * The schema history table used by flyway.
      */
-    private Table table;
-
-    /**
-     * Whether Flyway had to fall back to the old default table.
-     */
-    private boolean tableFallback;
+    private final Table table;
 
     /**
      * Connection with access to the database.
@@ -73,7 +67,7 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
     /**
      * The user invoking Flyway, for audit purposes.
      */
-    private String installedBy;
+    private final String installedBy;
 
     /**
      * Creates a new instance of the schema history table support.
@@ -83,11 +77,33 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
      * @param installedBy The user invoking Flyway, for audit purposes.
      */
     JdbcTableSchemaHistory(Database database, Table table, String installedBy) {
-        this.connection = database.getMainConnection();
         this.database = database;
-        this.table = table;
+        this.connection = database.getMainConnection();
+        this.jdbcTemplate = connection.getJdbcTemplate();
+        this.table = determineTable(table);
         this.installedBy = installedBy;
-        jdbcTemplate = connection.getJdbcTemplate();
+    }
+
+    /**
+     * Checks whether Flyway has to fallback to the old default table.
+     *
+     * @param table The currently configured table.
+     * @return The table to use.
+     */
+    private Table determineTable(Table table) {
+        // Ensure we are using the default table name before checking for the fallback table
+        if (table.getName().equals("flyway_schema_history")) {
+            Table fallbackTable = table.getSchema().getTable("schema_version");
+            if (fallbackTable.exists()) {
+                LOG.warn("Could not find schema history table " + table + ", but found " + fallbackTable + " instead." +
+                        " You are seeing this message because Flyway changed its default for flyway.table in" +
+                        " version 5.0.0 to flyway_schema_history and you are still relying on the old default (schema_version)." +
+                        " Set flyway.table=schema_version in your configuration to fix this." +
+                        " This fallback mechanism will be removed in Flyway 6.0.0.");
+                table = fallbackTable;
+            }
+        }
+        return table;
     }
 
     @Override
@@ -97,25 +113,12 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
 
     @Override
     public boolean exists() {
-        // Ensure we are using the default table name before checking for the fallback table
-        if (!tableFallback && table.getName().equals("flyway_schema_history")) {
-            Table fallbackTable = table.getSchema().getTable("schema_version");
-            if (fallbackTable.exists()) {
-                LOG.warn("Could not find schema history table " + table + ", but found " + fallbackTable + " instead." +
-                        " You are seeing this message because Flyway changed its default for flyway.table in" +
-                        " version 5.0.0 to flyway_schema_history and you are still relying on the old default (schema_version)." +
-                        " Set flyway.table=schema_version in your configuration to fix this." +
-                        " This fallback mechanism will be removed in Flyway 6.0.0.");
-                tableFallback = true;
-                table = fallbackTable;
-            }
-        }
+        connection.restoreOriginalState();
+
         return table.exists();
     }
 
-    /**
-     * Creates the metatable if it doesn't exist, upgrades it if it does.
-     */
+    @Override
     public void create() {
         int retries = 0;
         while (!exists()) {
@@ -127,8 +130,8 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
 
 
 
-                ).execute(connection.getJdbcTemplate());
-                LOG.debug("Schema History table " + table + " created.");
+                ).execute(jdbcTemplate);
+                LOG.debug("Created Schema History table: " + table);
             } catch (FlywayException e) {
                 if (++retries >= 10) {
                     throw e;
@@ -145,12 +148,16 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
 
     @Override
     public <T> T lock(Callable<T> callable) {
+        connection.restoreOriginalState();
+
         return connection.lock(table, callable);
     }
 
     @Override
-    protected void doAddAppliedMigration(MigrationVersion version, String description, MigrationType type, String script, Integer checksum, int executionTime, boolean success) {
-        connection.changeCurrentSchemaTo(table.getSchema());
+    protected void doAddAppliedMigration(int installedRank, MigrationVersion version, String description,
+                                         MigrationType type, String script, Integer checksum,
+                                         int executionTime, boolean success) {
+        connection.restoreOriginalState();
 
         // Lock again for databases with no DDL transactions to prevent implicit commits from triggering deadlocks
         // in highly concurrent environments
@@ -160,46 +167,29 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
 
         try {
             String versionStr = version == null ? null : version.toString();
-            int installedRank = type == MigrationType.SCHEMA ? 0 : calculateInstalledRank();
 
             jdbcTemplate.update(database.getInsertStatement(table),
                     installedRank, versionStr, description, type.name(), script, checksum, installedBy,
                     executionTime, success);
 
-            LOG.debug("Schema history table " + table + " successfully updated to reflect changes");
+            LOG.debug("Schema History table " + table + " successfully updated to reflect changes");
         } catch (SQLException e) {
             throw new FlywaySqlException("Unable to insert row for version '" + version + "' in Schema History table " + table, e);
         }
     }
 
-    /**
-     * Calculates the installed rank for the new migration to be inserted.
-     *
-     * @return The installed rank.
-     */
-    private int calculateInstalledRank() throws SQLException {
-        int currentMax = jdbcTemplate.queryForInt("SELECT MAX(" + database.quote("installed_rank") + ")"
-                + " FROM " + table);
-        return currentMax + 1;
-    }
-
     @Override
     public List<AppliedMigration> allAppliedMigrations() {
-        return findAppliedMigrations();
-    }
-
-    /**
-     * Retrieve the applied migrations from the schema history table.
-     *
-     * @param migrationTypes The specific migration types to look for. (Optional) None means find all migrations.
-     * @return The applied migrations.
-     */
-    private List<AppliedMigration> findAppliedMigrations(MigrationType... migrationTypes) {
         if (!exists()) {
             return new ArrayList<>();
         }
 
-        int minInstalledRank = cache.isEmpty() ? -1 : cache.getLast().getInstalledRank();
+        refreshCache();
+        return cache;
+    }
+
+    private void refreshCache() {
+        int maxCachedInstalledRank = cache.isEmpty() ? -1 : cache.getLast().getInstalledRank();
 
         String query = "SELECT " + database.quote("installed_rank")
                 + "," + database.quote("version")
@@ -212,22 +202,8 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
                 + "," + database.quote("execution_time")
                 + "," + database.quote("success")
                 + " FROM " + table
-                + " WHERE " + database.quote("installed_rank") + " > " + minInstalledRank;
-
-        if (migrationTypes.length > 0) {
-            query += " AND " + database.quote("type") + " IN (";
-            StringBuilder queryBuilder = new StringBuilder(query);
-            for (int i = 0; i < migrationTypes.length; i++) {
-                if (i > 0) {
-                    queryBuilder.append(",");
-                }
-                queryBuilder.append("'").append(migrationTypes[i]).append("'");
-            }
-            query = queryBuilder.toString();
-            query += ")";
-        }
-
-        query += " ORDER BY " + database.quote("installed_rank");
+                + " WHERE " + database.quote("installed_rank") + " > " + maxCachedInstalledRank
+                + " ORDER BY " + database.quote("installed_rank");
 
         try {
             cache.addAll(jdbcTemplate.query(query, new RowMapper<AppliedMigration>() {
@@ -251,7 +227,6 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
                     );
                 }
             }));
-            return cache;
         } catch (SQLException e) {
             throw new FlywaySqlException("Error while retrieving the list of applied migrations from Schema History table "
                     + table, e);
@@ -265,18 +240,20 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
             return;
         }
 
-        try {
-            int failedCount = jdbcTemplate.queryForInt("SELECT COUNT(*) FROM " + table
-                    + " WHERE " + database.quote("success") + "=" + database.getBooleanFalse());
-            if (failedCount == 0) {
-                LOG.info("Repair of failed migration in Schema History table " + table + " not necessary. No failed migration detected.");
-                return;
+        boolean failed = false;
+        List<AppliedMigration> appliedMigrations = allAppliedMigrations();
+        for (AppliedMigration appliedMigration : appliedMigrations) {
+            if (!appliedMigration.isSuccess()) {
+                failed = true;
             }
-        } catch (SQLException e) {
-            throw new FlywaySqlException("Unable to check the Schema History table " + table + " for failed migrations", e);
+        }
+        if (!failed) {
+            LOG.info("Repair of failed migration in Schema History table " + table + " not necessary. No failed migration detected.");
+            return;
         }
 
         try {
+            clearCache();
             jdbcTemplate.execute("DELETE FROM " + table
                     + " WHERE " + database.quote("success") + " = " + database.getBooleanFalse());
         } catch (SQLException e) {
@@ -285,67 +262,9 @@ public class JdbcTableSchemaHistory extends SchemaHistory {
     }
 
     @Override
-    public void addSchemasMarker(final Schema[] schemas) {
-        // Lock again for databases with no DDL transactions to prevent implicit commits from triggering deadlocks
-        // in highly concurrent environments
-        table.lock();
-
-        doAddSchemasMarker(schemas);
-    }
-
-    @Override
-    public boolean hasSchemasMarker() {
-        if (!table.exists()) {
-            return false;
-        }
-
-        try {
-            int count = jdbcTemplate.queryForInt(
-                    "SELECT COUNT(*) FROM " + table + " WHERE " + database.quote("type") + "='SCHEMA'");
-            return count > 0;
-        } catch (SQLException e) {
-            throw new FlywaySqlException("Unable to check whether the Schema History table " + table + " has a schema marker migration", e);
-        }
-    }
-
-    @Override
-    public boolean hasBaselineMarker() {
-        if (!table.exists()) {
-            return false;
-        }
-
-        try {
-            int count = jdbcTemplate.queryForInt(
-                    "SELECT COUNT(*) FROM " + table + " WHERE " + database.quote("type") + "='INIT' OR " + database.quote("type") + "='BASELINE'");
-            return count > 0;
-        } catch (SQLException e) {
-            throw new FlywaySqlException("Unable to check whether the Schema History table " + table + " has an baseline marker migration", e);
-        }
-    }
-
-    @Override
-    public AppliedMigration getBaselineMarker() {
-        List<AppliedMigration> appliedMigrations = findAppliedMigrations(MigrationType.BASELINE);
-        return appliedMigrations.isEmpty() ? null : appliedMigrations.get(0);
-    }
-
-    @Override
-    public boolean hasAppliedMigrations() {
-        if (!table.exists()) {
-            return false;
-        }
-
-        try {
-            int count = jdbcTemplate.queryForInt(
-                    "SELECT COUNT(*) FROM " + table + " WHERE " + database.quote("type") + " NOT IN ('SCHEMA', 'INIT', 'BASELINE')");
-            return count > 0;
-        } catch (SQLException e) {
-            throw new FlywaySqlException("Unable to check whether the Schema History table " + table + " has applied migrations", e);
-        }
-    }
-
-    @Override
     public void update(AppliedMigration appliedMigration, ResolvedMigration resolvedMigration) {
+        connection.restoreOriginalState();
+
         clearCache();
 
         MigrationVersion version = appliedMigration.getVersion();
